refactor(02-13): hoist NLog test initializer to Tests.Shared and fix NLog.config root cause (IN-02)
- Remove `concurrentWrites="true"` from `Sources/Libraries/ProjectV.Logging/NLog.config`. NLog 6 dropped that attribute, and combined with `throwConfigExceptions="true"` it caused the auto-load to throw `NLog.NLogConfigurationException` whenever any production type with a static `NLog.Logger` field was touched inside a test process.
- Hoist the empty-config `[ModuleInitializer]` body into `ProjectV.Tests.Shared.ForTests.TestModuleInitializer` so every test assembly that references Tests.Shared (all of them) inherits the suppression for free via the global-using of that namespace.
- Delete the 12 per-assembly `*ModuleInitializer.cs` / `*TestsInit.cs` files that used to carry an identical body (Appraisers, CommunicationWebService, Core, Crawlers, DataPipeline, Executors, InputProcessing, OmdbService, OutputProcessing, SteamService, TelegramBotWebService, TmdbService).
- Update three test-class XML-doc comments that previously `<see cref>`'d the deleted per-assembly initializer types (OutputManagerTests, InputManagerTests, CrawlersManagerTests) to point at the hoisted `ProjectV.Tests.Shared.ForTests.TestModuleInitializer` in prose form.

Build + Unit/Contract/Integration tests still 170 green; format clean.
