+++ Sources/Libraries/ProjectV.Logging/NLog.config
@@ -28,7 +28,6 @@
         <target name="logfile"
                 xsi:type="File"
                 fileName="${specialfolder:folder=CommonApplicationData}/ProjectV/logs/${shortdate:universalTime=false}.log"
-                concurrentWrites="true"
                 keepFileOpen="false"
                 header="${newline}---------------------------- Logging started on ${longdate:universalTime=false} ----------------------------"
                 layout="${var:logprefix} | ${logger} | ${replace-newlines:${message}} | ${onexception:${newline}${var:logprefix} | ${formatex}}"
